Button fragments (#36)
* change panel buttons to fragments (much faster now)

* add r2 and MAE to cellpose training plots

* add option to change cellpose hyperparameters

* add options for cellpose hyperparameter tuning

* add cellpose & densenet training plots to downloads

* reorganize data and model downloads

* redesign of mask editing side panel buttons

* test
