+++ requirements.txt
@@ -7,7 +7,7 @@ matplotlib
 Pillow
 scikit-learn
 scikit-image~=0.25
-streamlit
+streamlit==1.4.0
 tensorflow
 keras==3.10 # 3.11 does not work, which might be a bit restrictive
 torch>=2.5.1
